+++ idp/pkg/server/http/server.go
@@ -5,9 +5,8 @@ import (
 	"os"
 
 	"github.com/asim/go-micro/v3"
-
-	"github.com/owncloud/ocis/idp/pkg/crypto"
 	svc "github.com/owncloud/ocis/idp/pkg/service/v0"
+	pkgcrypto "github.com/owncloud/ocis/ocis-pkg/crypto"
 	"github.com/owncloud/ocis/ocis-pkg/middleware"
 	"github.com/owncloud/ocis/ocis-pkg/service/http"
 )
@@ -24,7 +23,7 @@ func Server(opts ...Option) (http.Service, error) {
 
 			if os.IsNotExist(certErr) || os.IsNotExist(keyErr) {
 				options.Logger.Info().Msgf("Generating certs")
-				if err := crypto.GenCert(options.Config.HTTP.TLSCert, options.Config.HTTP.TLSKey, options.Logger); err != nil {
+				if err := pkgcrypto.GenCert(options.Config.HTTP.TLSCert, options.Config.HTTP.TLSKey, options.Logger); err != nil {
 					options.Logger.Fatal().Err(err).Msg("Could not setup TLS")
 					os.Exit(1)
 				}
